backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
redis>=5.0
watchdog>=4.0
google-adk>=1.0
google-genai>=1.0
requests>=2.31
python-dotenv>=1.0
celery>=5.4.0
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
responses>=0.25
